+++ common-libs/internal-api-clients/src/types/typeBuilder.ts
@@ -4,13 +4,13 @@ import { RemoveEmpty, Simplify } from "./util"
 type BuildProperties<TProps extends DataObjectType['properties'], TObjects extends ParsedSpec['objects']> = {
   [key in keyof TProps as TProps[key]['isRequired'] extends true ? key : never]: BuildDataWithOptions<TProps[key], TObjects>
 } & {
-  [key in keyof TProps as TProps[key]['isRequired'] extends false ? key : never]?: BuildDataWithOptions<TProps[key], TObjects>
-}
+    [key in keyof TProps as TProps[key]['isRequired'] extends false ? key : never]?: BuildDataWithOptions<TProps[key], TObjects>
+  }
 
 type BuildAdditionalData<TIsAdditionalPropertiesAllowed extends boolean> =
   TIsAdditionalPropertiesAllowed extends true
-    ? Record<string, any>
-    : Record<never, never>
+  ? Record<string, any>
+  : Record<never, never>
 
 type BuildObjectData<TData extends DataObjectType, TObjects extends ParsedSpec['objects']> =
   BuildProperties<TData['properties'], TObjects> & BuildAdditionalData<TData['isAdditionalProperties']>
@@ -20,28 +20,28 @@ type BuildObjectDataWithUndefined<TData extends DataObjectType | undefined, TObj
 
 type BuildData<TData extends DataType, TObjects extends ParsedSpec['objects']> =
   TData extends DataRefType<infer URef>
-    ? BuildData<TObjects[URef], TObjects>
-    : TData extends DataPrimitiveType<infer UPrimitive>
-      ? UPrimitive
-      : TData extends DataRecordType
-        ? Record<string, any>
-        : TData extends DataArrayType
-          ? BuildDataWithOptions<TData['elementType'], TObjects>[]
-          : TData extends DataObjectType
-            ? BuildObjectData<TData, TObjects>
-            : TData extends DataAllOfType<infer UElementTypes>
-              ? never
-              : never
+  ? BuildData<TObjects[URef], TObjects>
+  : TData extends DataPrimitiveType<infer UPrimitive>
+  ? UPrimitive
+  : TData extends DataRecordType
+  ? Record<string, any>
+  : TData extends DataArrayType
+  ? BuildDataWithOptions<TData['elementType'], TObjects>[]
+  : TData extends DataObjectType
+  ? BuildObjectData<TData, TObjects>
+  : TData extends DataAllOfType<infer UElementTypes>
+  ? never
+  : never
 
 type BuildDataWithUndefined<TData extends DataType | undefined, TObjects extends ParsedSpec['objects']> =
   TData extends DataType ? BuildData<TData, TObjects> : undefined
 
 type BuildDataWithOptions<TData extends DataTypeWithOptions, TObjects extends ParsedSpec['objects']> =
   TData['isNullable'] extends true
-    ? BuildData<TData, TObjects> | null
-    : TData['isNullable'] extends false
-      ? BuildData<TData, TObjects>
-      : never
+  ? BuildData<TData, TObjects> | null
+  : TData['isNullable'] extends false
+  ? BuildData<TData, TObjects>
+  : never
 
 type BuildOperation<TOperation extends ParsedOperationSpec, TObjects extends ParsedSpec['objects']> = {
   pathParams: RemoveEmpty<BuildObjectDataWithUndefined<TOperation['pathParams'], TObjects>>
@@ -54,7 +54,7 @@ type BuildOperations<TOperations extends ParsedSpec['operations'], TObjects exte
   [key in keyof TOperations]: BuildOperation<TOperations[key], TObjects>
 }
 
-type BuildApiRawType<TParsedSpec extends ParsedSpec> = 
+type BuildApiRawType<TParsedSpec extends ParsedSpec> =
   BuildOperations<TParsedSpec['operations'], TParsedSpec['objects']>
 
 export type BuiltApiType = Record<string, {