+++ common-libs/internal-api-clients/CHANGELOG.md
@@ -1,3 +1,5 @@
+# release 1.0.0-beta.7 (2021-07-30)
+  * Created abstract `DidAuthApiService` and moved all Did auth related logic into it
 # release 1.0.0-beta.6 (2021-07-26)
   * Moved headers management to `helpers` module
   * Added `X-SDK-Version` header